A1.9.0
----
System
----
-Added Food icons
-Added Coin sounds
-Added give_gold Command
-updated voting system code
-small changes in player controllers
----
Game
------
-All* POI's are now randomly generated
-Added Blood Alter
-Added Hidden Rooms to Sector 1
-Added Slot Machine POI

+++ Assets/Prefabs/items/2-Items/Usble items/Effects/Effect Scripts/StatBuffItemEffect.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public class StatBuffItemEffect : UseableItemEffect
 {
     public int StrengthBuff;
-    public int AgilityBuff;
+    public float AgilityBuff;
     public int IntelligenceBuff;
     public int VitalityBuff;
     [HideIf("@!hasDuration")]
